requirements.txt
flask
google-generativeai
pymupdf
pytesseract
pillow
reportlab
